+++ crates/ruff_python_parser/resources/inline/err/params_multiple_slash_separator.py
@@ -1,2 +1,2 @@
 def foo(a, /, /, b): ...
-# def foo(a, /, b, c, /): ...
+def foo(a, /, b, c, /): ...
