+++ crates/ruff_python_parser/resources/inline/err/params_multiple_varargs.py
@@ -1,4 +1,4 @@
 def foo(a, *, *args, b): ...
 # def foo(a, *, b, c, *args): ...
 def foo(a, *args1, *args2, b): ...
-# def foo(a, *args1, b, c, *args2): ...
+def foo(a, *args1, b, c, *args2): ...
