+++ crates/ruff_python_parser/resources/invalid/expressions/lambda_default_parameters.py
@@ -1,2 +1,1 @@
-# TODO(micha): The offset of the generated error message is off by one.
 lambda a, b=20, c: 1
